requirements.txt
aiogram>=3.4.1
SQLAlchemy>=2.0.29
asyncpg>=0.29.0
alembic>=1.13.1
python-dotenv>=1.0.1
